New functions ft_put_pixel and ft_draw_circle, with some dirty extras and debug code.

+++ ft_draw_line.c
@@ -6,7 +6,7 @@
 /*   By: ngontjar <ngontjar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/20 18:25:34 by ngontjar          #+#    #+#             */
-/*   Updated: 2020/01/22 18:22:01 by ngontjar         ###   ########.fr       */
+/*   Updated: 2020/01/23 21:07:43 by ngontjar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ int					ft_draw_line(t_program *p, t_xy start, t_xy end, int color)
 	pos.y = start.y;
 	while (pixels--)
 	{
-		index = (int)((int)pos.y * p->buffer.line_bytes + pos.x);
+		index = ((int)pos.y * p->buffer.line_bytes + pos.x);
 		if (index < 0 || index > WIN_WIDTH * WIN_HEIGHT)
 		{
 			printf("ignoring pixel %.0f %.0f (%d)\n", pos.x, pos.y, index);
